requirements.txt
requests>=2.8,<2.9
aiohttp>=0.21,<0.22
aiohttp-jinja2>=0.6,<0.7
jinja2>=2.8,<2.9
path.py>=8.1,<8.2
Werkzeug>=0.11,<0.12
pywin32 > 1.0 ; sys_platform == 'win32'